Rename package 'frameworks' to 'infrastructure'
