+++ dj_link/__init__.py
@@ -1,4 +1,4 @@
 """A tool for linking two DataJoint tables located on different database servers."""
-from .frameworks.link import create_link as link
+from .infrastructure.link import create_link as link
 
 __all__ = ["link"]
