+++ dj_link/infrastructure/mixin.py
